refactor(#118): flattenPersonal 추출로 PATCH 중복 제거, ops[] 타입 명시
