+++ apps/api/src/app/api/mentee/basic-info/route.ts
@@ -1,14 +1,14 @@
 import { NextRequest, NextResponse } from "next/server";
-import { prisma } from "@plawcess/database";
+import { prisma, Prisma } from "@plawcess/database";
 import { getTokenFromCookie } from "@/lib/auth";
 import {
-  genderToLabel, labelToGender,
-  statusToLabel, labelToStatus,
-  militaryToLabel, labelToMilitary,
-  dateToLabel, labelToDate,
+  genderToLabel,
+  statusToLabel,
+  militaryToLabel,
+  dateToLabel,
   yearToLabel,
 } from "@/lib/labels";
-import { splitPayload, MENTEE_RECORD_FIELDS } from "@/lib/payload-split";
+import { splitPayload, MENTEE_RECORD_FIELDS, flattenPersonal, PersonalPatchInput } from "@/lib/payload-split";
 
 function getUserId(req: NextRequest): string | null {
   return getTokenFromCookie(req)?.user_id ?? null;
@@ -98,16 +98,7 @@ export async function PATCH(req: NextRequest) {
   const processYear = getProcessYear(req);
 
   let body: {
-    personal?: {
-      birthDate?: string;
-      gender?: string;
-      militaryStatus?: string;
-      major1?: string;
-      major2?: string;
-      admissionYear?: string;
-      academicStatus?: string;
-      graduationYear?: string;
-    };
+    personal?: PersonalPatchInput;
     admission?: {
       가?: { first?: string };
       나?: { first?: string };
@@ -121,24 +112,7 @@ export async function PATCH(req: NextRequest) {
   }
 
   // 1) 중첩 본문 → 평탄화 + DB 컬럼명 + 라벨 변환
-  const flat: Record<string, unknown> = {};
-  if (body.personal) {
-    const p = body.personal;
-    if (p.birthDate !== undefined) flat.birth_date = labelToDate(p.birthDate);
-    if (p.gender !== undefined) flat.gender = labelToGender(p.gender);
-    if (p.militaryStatus !== undefined) flat.military_status = labelToMilitary(p.militaryStatus);
-    if (p.major1 !== undefined) flat.undergrad_first_major = p.major1;
-    if (p.major2 !== undefined) flat.undergrad_second_major = p.major2;
-    if (p.admissionYear !== undefined) {
-      const n = parseInt(p.admissionYear, 10);
-      flat.undergrad_entry_year = isNaN(n) ? null : n;
-    }
-    if (p.graduationYear !== undefined) {
-      const n = parseInt(p.graduationYear, 10);
-      flat.undergrad_graduation_year = isNaN(n) ? null : n;
-    }
-    if (p.academicStatus !== undefined) flat.academic_status = labelToStatus(p.academicStatus);
-  }
+  const flat: Record<string, unknown> = body.personal ? flattenPersonal(body.personal) : {};
   if (body.admission) {
     const { 가: ga, 나: na, isSpecialAdmission } = body.admission;
     if (ga?.first !== undefined) flat.target_school_ga = ga.first || null;
@@ -156,7 +130,7 @@ export async function PATCH(req: NextRequest) {
   }
 
   // 4) 트랜잭션으로 두 테이블 동시 반영
-  const ops = [];
+  const ops: Prisma.PrismaPromise<unknown>[] = [];
   if (Object.keys(userData).length > 0) {
     ops.push(prisma.user.update({ where: { user_id: userId }, data: userData }));
   }
